Move `locate()` behind a feature flag
