+++ Cargo.toml
@@ -10,8 +10,11 @@ readme = "README.md"
 keywords = ["steam", "vdf", "appmanifest", "directory", "steamapps"]
 categories = ["os", "hardware-support", "filesystem", "accessibility"]
 
+# TODO: test more features in CI
+
 [features]
-default = []
+default = ["locate"]
+locate = ["locate_backend"]
 shortcuts_extras = ["crc"]
 
 [dependencies]
@@ -22,9 +25,9 @@ serde = { version = "1.0", features = ["derive"] }
 crc = { version = "3.0", optional = true }
 
 [target.'cfg(target_os="windows")'.dependencies]
-winreg = "0.51"
+locate_backend = { package = "winreg", version = "0.51", optional = true }
 [target.'cfg(not(target_os="windows"))'.dependencies]
-dirs = "5"
+locate_backend = { package = "dirs", version = "5", optional = true }
 
 [dev-dependencies]
 insta = { version = "1.34.0", features = ["redactions", "ron"] }
